+++ app/main/vo/__init__.py
@@ -1,0 +1,7 @@
+from app.main.vo.elasticsearch import ElasticsearchDocVO
+
+
+__all__ = [
+    ElasticsearchDocVO
+]
+