+++ app/main/vo/elasticsearch.py
@@ -1,0 +1,9 @@
+class ElasticsearchDocVO:
+
+    TIMESTAMP = 'timestamp'
+    UPDATED_AT = 'updated_at'
+    SOURCE = 'source'
+    CONTENT = 'content'
+    TOPICS = 'topics'
+    SENTIMENT = 'sentiment'
+